refactor(web-react): extract shared `useSelectionGridKeyboard` hook #DS-2304
Share tag-grid keyboard handling between `UNSTABLE_Picker` and `UNSTABLE_Combobox`.
