+++ packages/web-react/src/components/UNSTABLE_Picker/types.ts
@@ -1,5 +1,5 @@
-import { type FocusEvent, type KeyboardEvent, type ReactNode } from 'react';
-import type { SelectionMode } from '../../hooks';
+import { type ReactNode } from 'react';
+import type { SelectionGridRowProps, SelectionMode } from '../../hooks';
 import type {
   DropdownBaseProps,
   FillVariantDictionaryType,
@@ -17,14 +17,8 @@ import type {
 
 export type UnstablePickerSelectionMode = SelectionMode;
 
-/** Row props produced by `usePickerSelectionGridKeyboard` for roving tabindex and grid keys */
-export interface UnstablePickerSelectionGridRowProps {
-  tabIndex: 0 | -1;
-  onKeyDown: (event: KeyboardEvent<HTMLElement>) => void;
-  onFocusCapture: (event: FocusEvent<HTMLElement>) => void;
-  onBlurCapture: (event: FocusEvent<HTMLElement>) => void;
-  removeButtonTabIndex: 0 | -1;
-}
+/** Row props produced by `useSelectionGridKeyboard` for roving tabindex and grid keys */
+export type UnstablePickerSelectionGridRowProps = SelectionGridRowProps;
 
 export interface UnstablePickerTranslations {
   addButtonLabel?: string;
