+++ packages/web-react/src/hooks/index.ts
@@ -21,6 +21,7 @@ export * from './useLastActiveFocus';
 export * from './useLinkClick';
 export * from './useResizeObserver';
 export * from './useScrollControl';
+export * from './useSelectionGridKeyboard';
 export * from './useSelectionState';
 export * from './useSpacingStyle';
 export * from './useSymmetry';
